Bug Fix : Image upload issue
Problem Statement:
If you upload an image with the same exact title as of a previously uploaded image, it will get uploaded. But then, if you try to navigate to one of the images with the same title, the image uploader will display an error.
Fix/Resolution:
User will be able to navigate to the exact image (depending upon the image id stored at DB level) and no error/exception will be encountered at the front-end.