Bug Fix : Edit/Delete functionality
Problem Statement : 
After logging into the application, it is possible to edit/delete the image which is posted by some other user. This is a bug in the application. Now, fix this bug in the application, such that only the owner of the image can edit/delete the image.
Resolution/Fix:
We have incorporated the required changes at the code level to ensure that only authorized users are allowed to modify (edit/delete) the details for any specified image.